Separate coordinator in nibe heatpump
