+++ homeassistant/components/nibe_heatpump/button.py
@@ -11,7 +11,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import DOMAIN, LOGGER, Coordinator
+from .const import DOMAIN, LOGGER
+from .coordinator import Coordinator
 
 
 async def async_setup_entry(
